graph.R
For producing the plot of optimal acceptance rate versus dimensions. 
Although the range of values is better than before, they do not seem to converge the way we want them to. 
A probable reason is an inefficient search for optimal sigma value. We are finding sigma that minimizes the average asymptotic variance, But our vector is